* master: (26 commits)
  Edgecloud 3002 WebUI: ApplicationHA Policy Support (#469)
  Fixed a bug where no appInst, Cluster is displayed even though the cluster and app inst exist when entering from dev monitoring. (#470)
  EDGECLOUD-2933 Web: Automatic Cloudlet Onboarding Support Phase2  (#467)
  edgecloud 2997-  [Monitoring] Fix monitoring page layout when you expand/collapse mobiledgex left panel (#468)
  Webscoket URL was wrong (#466)
  EDGECLOUD-2134 Web UI - add Auto Scale Policy to Policies (#465)
  Edgecloud 2985,2943 (#463)
  Store last organization selection until login changes (#464)
  EDGECLOUD-2981 Need a UI knob to enable/disable health checking on the Application Access Ports (#462)
  Code Splitting Implementation for Console (#461)
  Edgecloud 2940 (Add Client Status for App Inst to Developer View and Mexadmin Monitoring) (#458)
  Organization audit redirects to side nav (#460)
  Edgecloud 2970 ,  edgecloud 2960 (#457)
  Using Syntax Highlighter for Json and Yaml View (#456)
  Implemented feature to filter the region of developer monitoring. (#455)
  EDGECLOUD-1961 Web UI - no error is given when MC is down (#454)
  EDGECLOUD-2949 [Developer-Monitoring] Can not scroll down monitoring (#451)
  Done (#453)
  EDGECLOUD-2057 Web UI - delete message for App Inst should include text about the cluster (#452)
  EDGECLOUD-2805 Web UI - Delete Cloudlet Pool should not be allowed if the pool contains an organization (#450)
  ...

# Conflicts:
#	.env
#	package.json
#	src/container/TimelineAuditView.js
#	src/services/model/endPointTypes.js
#	src/sites/siteFour/audits/auditLogs.js

+++ src/container/MexListView.js
@@ -218,27 +218,30 @@ class MexListView extends React.Component {
     will be changed to make it more generalize*/
     onActionClose = (action) => {
         let data = this.selectedRow;
-        switch (action.label) {
-            case 'Delete':
-                this.onWarning(action, 'delete', false, data)
-                break
-            case 'Upgrade':
-                this.onWarning(action, 'upgrade', false, data)
-                break;
-            case 'Refresh':
-                this.onWarning(action, 'refresh', false, data)
-                break;
-            case 'Power On':
-                this.onWarning(action, 'power on', false, data)
-                break;
-            case 'Power Off':
-                this.onWarning(action, 'power off', false, data)
-                break;
-            case 'Reboot':
-                this.onWarning(action, 'reboot', false, data)
-                break;
-            default:
-                action.onClick(action, data)
+        let valid = action.onClickInterept ? action.onClickInterept(action, data) : true
+        if (valid) {
+            switch (action.label) {
+                case 'Delete':
+                    this.onWarning(action, 'delete', false, data)
+                    break
+                case 'Upgrade':
+                    this.onWarning(action, 'upgrade', false, data)
+                    break;
+                case 'Refresh':
+                    this.onWarning(action, 'refresh', false, data)
+                    break;
+                case 'Power On':
+                    this.onWarning(action, 'power on', false, data)
+                    break;
+                case 'Power Off':
+                    this.onWarning(action, 'power off', false, data)
+                    break;
+                case 'Reboot':
+                    this.onWarning(action, 'reboot', false, data)
+                    break;
+                default:
+                    action.onClick(action, data)
+            }
         }
     }
